fix: apply bar graph api data
